requirements.txt
fastapi
psycopg2-binary
python-dotenv
python-multipart
pydantic
